+++ server/src/handlers/mod.rs
@@ -1,3 +1,4 @@
+pub mod auth;
 pub mod blog;
 pub mod category;
 pub mod picture;
