+++ server/src/routers/mod.rs
@@ -1,5 +1,6 @@
 use crate::AppState;
 use crate::core::journal::journal_request;
+use crate::routers::auth::create_auth_router;
 use crate::routers::blog::create_blog_router;
 use crate::routers::category::create_category_router;
 use crate::routers::picture::create_picture_router;
@@ -7,6 +8,7 @@ use crate::routers::user::create_user_router;
 use axum::{Extension, Router, middleware};
 use std::sync::Arc;
 
+mod auth;
 mod blog;
 mod category;
 mod picture;
@@ -14,6 +16,7 @@ mod user;
 
 pub fn create_app_router(state: Arc<AppState>) -> Router {
     Router::new()
+        .nest("/api", create_auth_router())
         .nest("/api", create_user_router())
         .nest("/api", create_category_router())
         .nest("/api", create_blog_router())
