+++ server/src/routers/user.rs
@@ -1,12 +1,10 @@
-use crate::handlers::user::{login, register, reset_password, token_login, verification_code};
+use crate::handlers::user::{decrease_user_permission, increase_user_permission, list_users};
 use axum::Router;
-use axum::routing::{get, post, put};
+use axum::routing::{get, put};
 
 pub fn create_user_router() -> Router {
     Router::new()
-        .route("/user/verification-code", post(verification_code))
-        .route("/user/register", post(register))
-        .route("/user/login", post(login))
-        .route("/user/jwt-login", get(token_login))
-        .route("/user/reset-password", put(reset_password))
+        .route("/user/list", get(list_users))
+        .route("/user/increase/{id}", put(increase_user_permission))
+        .route("/user/decrease/{id}", put(decrease_user_permission))
 }
